[#625] client/balance: remove intermediate wrapper
